Moved enums to new enums folder
