+++ src/main/java/org/icatproject/ids/PropertyHandler.java
@@ -22,10 +22,11 @@ import org.slf4j.LoggerFactory;
 
 import org.icatproject.ICAT;
 import org.icatproject.IcatException_Exception;
+import org.icatproject.ids.enums.CallType;
+import org.icatproject.ids.enums.StorageUnit;
 import org.icatproject.ids.plugin.ArchiveStorageInterface;
 import org.icatproject.ids.plugin.MainStorageInterface;
 import org.icatproject.ids.plugin.ZipMapperInterface;
-import org.icatproject.ids.v3.enums.CallType;
 import org.icatproject.utils.CheckedProperties;
 import org.icatproject.utils.CheckedProperties.CheckedPropertyException;
