+++ src/main/java/org/icatproject/ids/Tidier.java
@@ -24,13 +24,14 @@ import org.slf4j.LoggerFactory;
 import org.icatproject.Datafile;
 import org.icatproject.Dataset;
 import org.icatproject.IcatException_Exception;
+import org.icatproject.ids.enums.DeferredOp;
+import org.icatproject.ids.enums.StorageUnit;
 import org.icatproject.ids.exceptions.InsufficientPrivilegesException;
 import org.icatproject.ids.exceptions.InternalException;
 import org.icatproject.ids.plugin.DfInfo;
 import org.icatproject.ids.plugin.DsInfo;
 import org.icatproject.ids.plugin.MainStorageInterface;
 import org.icatproject.ids.v3.FiniteStateMachine.FiniteStateMachine;
-import org.icatproject.ids.v3.enums.DeferredOp;
 import org.icatproject.ids.v3.helper.LocationHelper;
 import org.icatproject.ids.v3.models.DataFileInfo;
 import org.icatproject.ids.v3.models.DataSetInfo;
