+++ src/main/java/org/icatproject/ids/enums/StorageUnit.java
@@ -1,4 +1,4 @@
-package org.icatproject.ids;
+package org.icatproject.ids.enums;
 
 public enum StorageUnit {
     DATASET, DATAFILE
