+++ src/main/java/org/icatproject/ids/v3/DataSelectionFactory.java
@@ -20,13 +20,13 @@ import org.icatproject.icat.client.IcatException;
 import org.icatproject.icat.client.Session;
 import org.icatproject.ids.IcatReader;
 import org.icatproject.ids.PropertyHandler;
-import org.icatproject.ids.StorageUnit;
+import org.icatproject.ids.enums.RequestType;
+import org.icatproject.ids.enums.StorageUnit;
 import org.icatproject.ids.exceptions.BadRequestException;
 import org.icatproject.ids.exceptions.InsufficientPrivilegesException;
 import org.icatproject.ids.exceptions.InternalException;
 import org.icatproject.ids.exceptions.NotFoundException;
 import org.icatproject.ids.exceptions.NotImplementedException;
-import org.icatproject.ids.v3.enums.RequestType;
 import org.icatproject.ids.v3.helper.LocationHelper;
 import org.icatproject.ids.v3.models.DataFileInfo;
 import org.icatproject.ids.v3.models.DataInfoBase;
