+++ src/main/java/org/icatproject/ids/v3/FiniteStateMachine/FiniteStateMachine.java
@@ -22,10 +22,10 @@ import org.slf4j.LoggerFactory;
 import org.icatproject.ids.IcatReader;
 import org.icatproject.ids.LockManager;
 import org.icatproject.ids.PropertyHandler;
-import org.icatproject.ids.StorageUnit;
 import org.icatproject.ids.LockManager.LockInfo;
+import org.icatproject.ids.enums.DeferredOp;
+import org.icatproject.ids.enums.StorageUnit;
 import org.icatproject.ids.exceptions.InternalException;
-import org.icatproject.ids.v3.enums.DeferredOp;
 import org.icatproject.ids.v3.models.DataInfoBase;
 
 public abstract class FiniteStateMachine {
