+++ src/main/java/org/icatproject/ids/v3/FiniteStateMachine/FiniteStateMachineForSingleLevelStorage.java
@@ -2,8 +2,8 @@ package org.icatproject.ids.v3.FiniteStateMachine;
 
 import org.icatproject.ids.IcatReader;
 import org.icatproject.ids.LockManager;
+import org.icatproject.ids.enums.DeferredOp;
 import org.icatproject.ids.exceptions.InternalException;
-import org.icatproject.ids.v3.enums.DeferredOp;
 import org.icatproject.ids.v3.models.DataInfoBase;
 
 import jakarta.json.stream.JsonGenerator;
