+++ src/main/java/org/icatproject/ids/v3/RequestHandlerService.java
@@ -5,6 +5,7 @@ import java.nio.file.Path;
 import java.util.HashMap;
 
 import org.icatproject.ids.PropertyHandler;
+import org.icatproject.ids.enums.RequestType;
 import org.icatproject.ids.exceptions.BadRequestException;
 import org.icatproject.ids.exceptions.DataNotOnlineException;
 import org.icatproject.ids.exceptions.InsufficientPrivilegesException;
@@ -12,7 +13,6 @@ import org.icatproject.ids.exceptions.InternalException;
 import org.icatproject.ids.exceptions.NotFoundException;
 import org.icatproject.ids.exceptions.NotImplementedException;
 import org.icatproject.ids.plugin.ArchiveStorageInterface;
-import org.icatproject.ids.v3.enums.RequestType;
 import org.icatproject.ids.v3.handlers.ArchiveHandler;
 import org.icatproject.ids.v3.handlers.DeleteHandler;
 import org.icatproject.ids.v3.handlers.GetDataFileIdsHandler;
