+++ src/main/java/org/icatproject/ids/v3/ServiceProvider.java
@@ -7,9 +7,9 @@ import org.icatproject.ids.IcatReader;
 import org.icatproject.ids.LockManager;
 import org.icatproject.ids.PropertyHandler;
 import org.icatproject.ids.Transmitter;
+import org.icatproject.ids.enums.CallType;
 import org.icatproject.ids.plugin.MainStorageInterface;
 import org.icatproject.ids.v3.FiniteStateMachine.FiniteStateMachine;
-import org.icatproject.ids.v3.enums.CallType;
 
 /**
  * This class serves the developer with multiple services.
