+++ src/main/java/org/icatproject/ids/v3/UnfinishedWorkService.java
@@ -9,10 +9,10 @@ import org.icatproject.Datafile;
 import org.icatproject.Dataset;
 import org.icatproject.IcatExceptionType;
 import org.icatproject.IcatException_Exception;
-import org.icatproject.ids.StorageUnit;
+import org.icatproject.ids.enums.DeferredOp;
+import org.icatproject.ids.enums.StorageUnit;
 import org.icatproject.ids.exceptions.InsufficientPrivilegesException;
 import org.icatproject.ids.exceptions.InternalException;
-import org.icatproject.ids.v3.enums.DeferredOp;
 import org.icatproject.ids.v3.helper.LocationHelper;
 import org.icatproject.ids.v3.models.DataFileInfo;
 import org.icatproject.ids.v3.models.DataInfoBase;
