+++ src/main/java/org/icatproject/ids/v3/handlers/PrepareDataHandler.java
@@ -11,7 +11,9 @@ import java.util.Set;
 import java.util.UUID;
 
 import org.icatproject.IcatException_Exception;
-import org.icatproject.ids.StorageUnit;
+import org.icatproject.ids.enums.CallType;
+import org.icatproject.ids.enums.RequestType;
+import org.icatproject.ids.enums.StorageUnit;
 import org.icatproject.ids.exceptions.BadRequestException;
 import org.icatproject.ids.exceptions.DataNotOnlineException;
 import org.icatproject.ids.exceptions.InsufficientPrivilegesException;
@@ -21,8 +23,6 @@ import org.icatproject.ids.exceptions.NotImplementedException;
 import org.icatproject.ids.v3.DataSelectionV3Base;
 import org.icatproject.ids.v3.RequestHandlerBase;
 import org.icatproject.ids.v3.ServiceProvider;
-import org.icatproject.ids.v3.enums.CallType;
-import org.icatproject.ids.v3.enums.RequestType;
 import org.icatproject.ids.v3.models.DataInfoBase;
 import org.icatproject.ids.v3.models.ValueContainer;
