+++ src/main/java/org/icatproject/ids/v3/handlers/RestoreHandler.java
@@ -4,7 +4,10 @@ import java.io.ByteArrayOutputStream;
 import java.util.HashMap;
 
 import org.icatproject.IcatException_Exception;
-import org.icatproject.ids.StorageUnit;
+import org.icatproject.ids.enums.CallType;
+import org.icatproject.ids.enums.DeferredOp;
+import org.icatproject.ids.enums.RequestType;
+import org.icatproject.ids.enums.StorageUnit;
 import org.icatproject.ids.exceptions.BadRequestException;
 import org.icatproject.ids.exceptions.DataNotOnlineException;
 import org.icatproject.ids.exceptions.InsufficientPrivilegesException;
@@ -14,9 +17,6 @@ import org.icatproject.ids.exceptions.NotImplementedException;
 import org.icatproject.ids.v3.DataSelectionV3Base;
 import org.icatproject.ids.v3.RequestHandlerBase;
 import org.icatproject.ids.v3.ServiceProvider;
-import org.icatproject.ids.v3.enums.CallType;
-import org.icatproject.ids.v3.enums.DeferredOp;
-import org.icatproject.ids.v3.enums.RequestType;
 import org.icatproject.ids.v3.models.ValueContainer;
 
 import jakarta.json.Json;
